refactor: moves components and hooks into the correct folder structure
