+++ components/MDX/SearchPage/index.tsx
